# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VentWall API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ReactionCounts {
  heard: number;
  iFeelThat: number;
  sameWave: number;
  notAlone: number;
  letItBurn: number;
  keepGoing: number;
  neededThis: number;
}

export interface Post {
  id: number;
  content: string;
  /** @nullable */
  emotionTag?: string | null;
  /** @nullable */
  authorName?: string | null;
  isAnonymous: boolean;
  expiresAt: string;
  createdAt: string;
  reactionCounts: ReactionCounts;
  reportCount: number;
  isRemoved: boolean;
  isAiReflection: boolean;
  isDailyPrompt: boolean;
}

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
  emotionTag?: string;
  authorName?: string;
  isAnonymous: boolean;
}

export type EchoInputReactionType = typeof EchoInputReactionType[keyof typeof EchoInputReactionType];


export const EchoInputReactionType = {
  heard: 'heard',
  iFeelThat: 'iFeelThat',
  sameWave: 'sameWave',
  notAlone: 'notAlone',
  letItBurn: 'letItBurn',
  keepGoing: 'keepGoing',
  neededThis: 'neededThis',
} as const;

export interface EchoInput {
  reactionType: EchoInputReactionType;
}

export interface ReportInput {
  reason: string;
  details?: string;
}

export interface Report {
  id: number;
  postId: number;
  reason: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface ContentCheck {
  content: string;
}

export type SafetyResultFlagType = typeof SafetyResultFlagType[keyof typeof SafetyResultFlagType];


export const SafetyResultFlagType = {
  none: 'none',
  selfHarm: 'selfHarm',
  threat: 'threat',
  blocked: 'blocked',
  harassment: 'harassment',
  spam: 'spam',
} as const;

export type SafetyResultSeverity = typeof SafetyResultSeverity[keyof typeof SafetyResultSeverity];


export const SafetyResultSeverity = {
  none: 'none',
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface SafetyResult {
  flagType: SafetyResultFlagType;
  severity: SafetyResultSeverity;
  allowed: boolean;
  /** @nullable */
  message: string | null;
}

export interface FeedSummary {
  activePostCount: number;
  totalReactions: number;
  /** @nullable */
  topEmotion: string | null;
}

export interface AdminPost {
  id: number;
  content: string;
  /** @nullable */
  emotionTag?: string | null;
  /** @nullable */
  authorName?: string | null;
  isAnonymous: boolean;
  expiresAt: string;
  createdAt: string;
  reactionCounts: ReactionCounts;
  reportCount: number;
  isRemoved: boolean;
  isFlagged: boolean;
  /** @nullable */
  moderationNote?: string | null;
}

export interface RateLimitResponse {
  error: string;
  retryAfterSeconds: number;
}

export interface DailyPrompt {
  prompt: string;
  date: string;
}

export type ModerationActionAction = typeof ModerationActionAction[keyof typeof ModerationActionAction];


export const ModerationActionAction = {
  remove: 'remove',
  flag: 'flag',
  unflag: 'unflag',
  restore: 'restore',
} as const;

export interface ModerationAction {
  action: ModerationActionAction;
  note?: string;
}

export type ListPostsParams = {
emotionTag?: string;
/**
 * Filter to posts with no emotion tag
 */
noTag?: boolean;
cursor?: number;
limit?: number;
};

export type ListAdminPostsParams = {
flagged?: boolean;
cursor?: number;
};

export type ListAdminReportsParams = {
cursor?: number;
};