# lib/api-zod/src/generated/types/echoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VentWall API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EchoInputReactionType } from './echoInputReactionType';

export interface EchoInput {
  reactionType: EchoInputReactionType;
}