# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VentWall API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
  emotionTag?: string;
  authorName?: string;
  isAnonymous: boolean;
}