# lib/api-zod/src/generated/types/safetyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VentWall API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SafetyResultFlagType } from './safetyResultFlagType';
import type { SafetyResultSeverity } from './safetyResultSeverity';

export interface SafetyResult {
  flagType: SafetyResultFlagType;
  severity: SafetyResultSeverity;
  allowed: boolean;
  /** @nullable */
  message: string | null;
}